refactor(lotto): LottoBunch에 집중된로직 분산화
