+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,8 @@
 package lotto
 
+import lotto.domain.factory.ManualLottoFactory
+import lotto.domain.factory.RandomLottoFactory
+
 fun main() {
-    LottoController().runLotto()
+    LottoController(manualLottoFactory = ManualLottoFactory(), autoLottoFactory = RandomLottoFactory()).runLotto()
 }
